+++ src/main/java/com/emc/mongoose/Main.java
@@ -82,10 +82,10 @@ public final class Main {
 				}
 				// init the metrics manager
 				final MetricsManager metricsMgr = new MetricsManagerImpl(ServiceTaskExecutor.INSTANCE);
+				final int port = configWithArgs.intVal("run-port");
 				if(configWithArgs.boolVal("run-node")) {
 					runNode(configWithArgs, extensions, metricsMgr);
 				} else {
-					final int port = configWithArgs.intVal("run-port");
 					final Server server = new Server(port);
 					addMetricsService(server);
 					server.start();
