+++ src/main/java/com/emc/mongoose/metrics/MetricsManagerImpl.java
@@ -12,20 +12,14 @@ import com.emc.mongoose.metrics.context.MetricsContext;
 import com.emc.mongoose.metrics.snapshot.AllMetricsSnapshot;
 import com.emc.mongoose.metrics.snapshot.ConcurrencyMetricSnapshot;
 import com.emc.mongoose.metrics.util.PrometheusMetricsExporter;
-import static com.emc.mongoose.Constants.KEY_CLASS_NAME;
-import static com.emc.mongoose.Constants.KEY_STEP_ID;
-import static com.emc.mongoose.metrics.MetricsConstants.METRIC_LABELS;
-
+import com.emc.mongoose.metrics.util.PrometheusMetricsExporterImpl;
 import com.github.akurilov.fiber4j.ExclusiveFiberBase;
 import com.github.akurilov.fiber4j.Fiber;
 import com.github.akurilov.fiber4j.FibersExecutor;
-
+import io.prometheus.client.Collector;
 import io.prometheus.client.CollectorRegistry;
-
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.ThreadContext;
-import static org.apache.logging.log4j.CloseableThreadContext.Instance;
-import static org.apache.logging.log4j.CloseableThreadContext.put;
 
 import java.util.Collections;
 import java.util.ConcurrentModificationException;
@@ -38,12 +32,18 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
+import static com.emc.mongoose.Constants.KEY_CLASS_NAME;
+import static com.emc.mongoose.Constants.KEY_STEP_ID;
+import static com.emc.mongoose.metrics.MetricsConstants.METRIC_LABELS;
+import static org.apache.logging.log4j.CloseableThreadContext.Instance;
+import static org.apache.logging.log4j.CloseableThreadContext.put;
+
 /**
  Created by kurila on 18.05.17.
  */
 public class MetricsManagerImpl
-extends ExclusiveFiberBase
-implements MetricsManager {
+	extends ExclusiveFiberBase
+	implements MetricsManager {
 
 	private static final String CLS_NAME = MetricsManagerImpl.class.getSimpleName();
 	private final Set<MetricsContext> allMetrics = new ConcurrentSkipListSet<>();
@@ -104,7 +104,7 @@ implements MetricsManager {
 					}
 				}
 				// console output
-				if(!selectedMetrics.isEmpty()) {
+				if(! selectedMetrics.isEmpty()) {
 					Loggers.METRICS_STD_OUT.info(new MetricsAsciiTableLogMessage(selectedMetrics));
 					selectedMetrics.clear();
 				}
@@ -118,7 +118,7 @@ implements MetricsManager {
 	}
 
 	public void startIfNotStarted() {
-		if(!isStarted()) {
+		if(! isStarted()) {
 			super.start();
 			Loggers.MSG.debug("Started the metrics manager fiber");
 		}
@@ -140,8 +140,9 @@ implements MetricsManager {
 				};
 				distributedMetrics.put(
 					distributedMetricsCtx,
-					new PrometheusMetricsExporter(distributedMetricsCtx)
+					new PrometheusMetricsExporterImpl(distributedMetricsCtx)
 						.labels(METRIC_LABELS, labelValues)
+						.quantiles(distributedMetricsCtx.quantileValues())
 						.register()
 				);
 			}
@@ -162,7 +163,7 @@ implements MetricsManager {
 		) {
 			if(allMetrics.remove(metricsCtx)) {
 				try {
-					if(!outputLock.tryLock(Fiber.WARN_DURATION_LIMIT_NANOS, TimeUnit.NANOSECONDS)) {
+					if(! outputLock.tryLock(Fiber.WARN_DURATION_LIMIT_NANOS, TimeUnit.NANOSECONDS)) {
 						Loggers.ERR.warn(
 							"Acquire lock timeout while unregistering the metrics context \"{}\"", metricsCtx
 						);
@@ -196,7 +197,7 @@ implements MetricsManager {
 						Loggers.METRICS_STD_OUT.info(new StepResultsMetricsLogMessage(distributedMetricsCtx));
 						final PrometheusMetricsExporter exporter = distributedMetrics.remove(distributedMetricsCtx);
 						if(exporter != null) {
-							CollectorRegistry.defaultRegistry.unregister(exporter);
+							CollectorRegistry.defaultRegistry.unregister((Collector) exporter);
 						}
 					}
 				} catch(final InterruptedException e) {
