+++ src/main/java/com/emc/mongoose/metrics/context/DistributedMetricsContext.java
@@ -2,10 +2,14 @@ package com.emc.mongoose.metrics.context;
 
 import com.emc.mongoose.metrics.snapshot.DistributedAllMetricsSnapshot;
 
+import java.util.List;
+
 public interface DistributedMetricsContext<S extends DistributedAllMetricsSnapshot>
 extends MetricsContext<S> {
 
 	int nodeCount();
+
+	List<Double> quantileValues();
 
 	S lastSnapshot();
 }
