+++ src/main/resources/config/defaults.json
@@ -1,10 +1,10 @@
 {
-	"item": {
-		"data": {
-			"input": {
+	"item" : {
+		"data" : {
+			"input" : {
 				// The source file for the content/payload generation
-				"file": null,
-				"layer": {
+				"file" : null,
+				"layer" : {
 					// The maximum count of the data "layers" to be cached into the memory
 					"cache": 16,
 					// The size of the content source ring buffer
@@ -157,7 +157,12 @@
 			},
 			// The concurrency threshold to enable intermediate statistics calculation,
 			// 0 means no threshold
-			"threshold": 0
+			"threshold" : 0,
+			"quantiles" : [
+				0.25,
+				0.5,
+				0.75
+			]
 		}
 	},
 	"run": {
